Get accessibility data for particular year and table number

The DfT's Journey Time Statistics are outlined here on the [gov.uk website](https://www.gov.uk/government/statistical-data-sets/journey-time-statistics-data-tables-jts).

get_jts_data(table, year = NULL, u_csv = NULL, skip = 6)

Arguments

table
The title of the table, e.g. "jts0501"

year
The year, e.g. 2017

u_csv
The csv url (not usually needed)

skip
How many lines to skip before the data starts (6 by default)

Details

The function uses a data frame of existing tables, created by the script 'accessibility_tables.R' in the data-raw folder.

The tables starting JTS01 to JTS03 provide national overview data. The tables JTS0401 to JTS0409 provide data at the Local Authority level. The tables JTS0501 to JTS0509 provide the same data at the LSOA level. The tables beginning JTS09 provide data on accessibility to transport hubs. And the tables beginning JTS10 contain add other variables.

Data is provided every year from 2014 to 2019 in many cases

Examples

jts_tables

#> # A tibble: 223 x 5
#>    table_code csv_file_name  table_title         table_url        csv_url       
#>    <chr>      <chr>          <chr>               <chr>            <chr>         
#>  1 jts0101    jts0101-2014.… Average minimum tr… https://assets.… https://githu…
#>  2 jts0101    jts0101-2015.… Average minimum tr… https://assets.… https://githu…
#>  3 jts0101    jts0101-2016.… Average minimum tr… https://assets.… https://githu…
#>  4 jts0101    jts0101-2017.… Average minimum tr… https://assets.… https://githu…
#>  5 jts0101    jts0101-JTS01… Average minimum tr… https://assets.… https://githu…
#>  6 jts0102    jts0102-2014.… Average minimum tr… https://assets.… https://githu…
#>  7 jts0102    jts0102-2015.… Average minimum tr… https://assets.… https://githu…
#>  8 jts0102    jts0102-2016.… Average minimum tr… https://assets.… https://githu…
#>  9 jts0102    jts0102-2017.… Average minimum tr… https://assets.… https://githu…
#> 10 jts0103    jts0103-2014.… Average minimum tr… https://assets.… https://githu…
#> # … with 213 more rows

jts_tables$table_title

#>   [1] "Average minimum travel time to reach the nearest key services by mode of travel: England"                                         
#>   [2] "Average minimum travel time to reach the nearest key services by mode of travel: England"                                         
#>   [3] "Average minimum travel time to reach the nearest key services by mode of travel: England"                                         
#>   [4] "Average minimum travel time to reach the nearest key services by mode of travel: England"                                         
#>   [5] "Average minimum travel time to reach the nearest key services by mode of travel: England"                                         
#>   [6] "Average minimum travel time to reach the nearest key services by mode of travel, rural and urban areas: England"                  
#>   [7] "Average minimum travel time to reach the nearest key services by mode of travel, rural and urban areas: England"                  
#>   [8] "Average minimum travel time to reach the nearest key services by mode of travel, rural and urban areas: England"                  
#>   [9] "Average minimum travel time to reach the nearest key services by mode of travel, rural and urban areas: England"                  
#>  [10] "Average minimum travel time to reach the nearest key services by public transport or walking, region: England"                    
#>  [11] "Average minimum travel time to reach the nearest key services by public transport or walking, region: England"                    
#>  [12] "Average minimum travel time to reach the nearest key services by public transport or walking, region: England"                    
#>  [13] "Average minimum travel time to reach the nearest key services by public transport or walking, region: England"                    
#>  [14] "Average minimum travel time to reach the nearest key services by public transport or walking, region: England"                    
#>  [15] "Average minimum travel time to reach the nearest key services by mode of travel, local authority: England"                        
#>  [16] "Average minimum travel time to reach the nearest key services by mode of travel, local authority: England"                        
#>  [17] "Average minimum travel time to reach the nearest key services by mode of travel, local authority: England"                        
#>  [18] "Average minimum travel time to reach the nearest key services by mode of travel, local authority: England"                        
#>  [19] "Service users with access to key services within selected journey times by mode of travel: England"                               
#>  [20] "Service users with access to key services within selected journey times by mode of travel: England"                               
#>  [21] "Service users with access to key services within selected journey times by mode of travel: England"                               
#>  [22] "Service users with access to key services within selected journey times by mode of travel: England"                               
#>  [23] "Service users with access to key services within selected journey times by mode of travel, rural and urban areas: England "       
#>  [24] "Service users with access to key services within selected journey times by mode of travel, rural and urban areas: England "       
#>  [25] "Service users with access to key services within selected journey times by mode of travel, rural and urban areas: England "       
#>  [26] "Service users with access to key services within selected journey times by mode of travel, rural and urban areas: England "       
#>  [27] "Service users with access to key services within selected journey times by public transport or walking, in regions: England"      
#>  [28] "Service users with access to key services within selected journey times by public transport or walking, in regions: England"      
#>  [29] "Service users with access to key services within selected journey times by public transport or walking, in regions: England"      
#>  [30] "Service users with access to key services within selected journey times by public transport or walking, in regions: England"      
#>  [31] "Service users with access to key services within selected journey times by cycle, in regions: England"                            
#>  [32] "Service users with access to key services within selected journey times by cycle, in regions: England"                            
#>  [33] "Service users with access to key services within selected journey times by cycle, in regions: England"                            
#>  [34] "Service users with access to key services within selected journey times by cycle, in regions: England"                            
#>  [35] "Service users with access to key services within selected journey times by car, in regions: England"                              
#>  [36] "Service users with access to key services within selected journey times by car, in regions: England"                              
#>  [37] "Service users with access to key services within selected journey times by car, in regions: England"                              
#>  [38] "Service users with access to key services within selected journey times by car, in regions: England"                              
#>  [39] "Number of sites available within selected journey times by key services, mode of travel: England "                                
#>  [40] "Number of sites available within selected journey times by key services, mode of travel: England "                                
#>  [41] "Number of sites available within selected journey times by key services, mode of travel: England "                                
#>  [42] "Number of sites available within selected journey times by key services, mode of travel: England "                                
#>  [43] "Number of sites available within selected journey times by key services, mode of travel, rural and urban areas: England"          
#>  [44] "Number of sites available within selected journey times by key services, mode of travel, rural and urban areas: England"          
#>  [45] "Number of sites available within selected journey times by key services, mode of travel, rural and urban areas: England"          
#>  [46] "Number of sites available within selected journey times by key services, mode of travel, rural and urban areas: England"          
#>  [47] "Number of sites available within selected journey times by key services, public transport or walking, in regions: England"        
#>  [48] "Number of sites available within selected journey times by key services, public transport or walking, in regions: England"        
#>  [49] "Number of sites available within selected journey times by key services, public transport or walking, in regions: England"        
#>  [50] "Number of sites available within selected journey times by key services, public transport or walking, in regions: England"        
#>  [51] "Number of sites available within selected journey times by key services, cycle, in regions: England"                              
#>  [52] "Number of sites available within selected journey times by key services, cycle, in regions: England"                              
#>  [53] "Number of sites available within selected journey times by key services, cycle, in regions: England"                              
#>  [54] "Number of sites available within selected journey times by key services, cycle, in regions: England"                              
#>  [55] "Number of sites available within selected journey times by key services, car in regions: England"                                 
#>  [56] "Number of sites available within selected journey times by key services, car in regions: England"                                 
#>  [57] "Number of sites available within selected journey times by key services, car in regions: England"                                 
#>  [58] "Number of sites available within selected journey times by key services, car in regions: England"                                 
#>  [59] "Travel time, destination and origin indicators for Employment centres by mode of travel and local authority, England "            
#>  [60] "Travel time, destination and origin indicators for Employment centres by mode of travel and local authority, England "            
#>  [61] "Travel time, destination and origin indicators for Employment centres by mode of travel and local authority, England "            
#>  [62] "Travel time, destination and origin indicators for Employment centres by mode of travel and local authority, England "            
#>  [63] "Travel time, destination and origin indicators for Employment centres by mode of travel and local authority, England "            
#>  [64] "Travel time, destination and origin indicators for Primary schools by mode of travel and local authority, England"                
#>  [65] "Travel time, destination and origin indicators for Primary schools by mode of travel and local authority, England"                
#>  [66] "Travel time, destination and origin indicators for Primary schools by mode of travel and local authority, England"                
#>  [67] "Travel time, destination and origin indicators for Primary schools by mode of travel and local authority, England"                
#>  [68] "Travel time, destination and origin indicators for Primary schools by mode of travel and local authority, England"                
#>  [69] "Travel time, destination and origin indicators for Secondary schools by mode of travel and local authority, England"              
#>  [70] "Travel time, destination and origin indicators for Secondary schools by mode of travel and local authority, England"              
#>  [71] "Travel time, destination and origin indicators for Secondary schools by mode of travel and local authority, England"              
#>  [72] "Travel time, destination and origin indicators for Secondary schools by mode of travel and local authority, England"              
#>  [73] "Travel time, destination and origin indicators for Secondary schools by mode of travel and local authority, England"              
#>  [74] "Travel time, destination and origin indicators for Further education by mode of travel and local authority, England"              
#>  [75] "Travel time, destination and origin indicators for Further education by mode of travel and local authority, England"              
#>  [76] "Travel time, destination and origin indicators for Further education by mode of travel and local authority, England"              
#>  [77] "Travel time, destination and origin indicators for Further education by mode of travel and local authority, England"              
#>  [78] "Travel time, destination and origin indicators for Further education by mode of travel and local authority, England"              
#>  [79] "Travel time, destination and origin indicators for GPs by mode of travel and local authority, England"                            
#>  [80] "Travel time, destination and origin indicators for GPs by mode of travel and local authority, England"                            
#>  [81] "Travel time, destination and origin indicators for GPs by mode of travel and local authority, England"                            
#>  [82] "Travel time, destination and origin indicators for GPs by mode of travel and local authority, England"                            
#>  [83] "Travel time, destination and origin indicators for GPs by mode of travel and local authority, England"                            
#>  [84] "Travel time, destination and origin indicators for Hospitals by mode of travel and local authority, England"                      
#>  [85] "Travel time, destination and origin indicators for Hospitals by mode of travel and local authority, England"                      
#>  [86] "Travel time, destination and origin indicators for Hospitals by mode of travel and local authority, England"                      
#>  [87] "Travel time, destination and origin indicators for Hospitals by mode of travel and local authority, England"                      
#>  [88] "Travel time, destination and origin indicators for Hospitals by mode of travel and local authority, England"                      
#>  [89] "Travel time, destination and origin indicators for Food stores by mode of travel and local authority, England"                    
#>  [90] "Travel time, destination and origin indicators for Food stores by mode of travel and local authority, England"                    
#>  [91] "Travel time, destination and origin indicators for Food stores by mode of travel and local authority, England"                    
#>  [92] "Travel time, destination and origin indicators for Food stores by mode of travel and local authority, England"                    
#>  [93] "Travel time, destination and origin indicators for Food stores by mode of travel and local authority, England"                    
#>  [94] "Travel time, destination and origin indicators for Town centres by mode of travel and local authority, England"                   
#>  [95] "Travel time, destination and origin indicators for Town centres by mode of travel and local authority, England"                   
#>  [96] "Travel time, destination and origin indicators for Town centres by mode of travel and local authority, England"                   
#>  [97] "Travel time, destination and origin indicators for Town centres by mode of travel and local authority, England"                   
#>  [98] "Travel time, destination and origin indicators for Town centres by mode of travel and local authority, England"                   
#>  [99] "Travel time, destination and origin indicators to Pharmacy by cycle and car, local authority, England"                            
#> [100] "Travel time, destination and origin indicators to Pharmacy by cycle and car, local authority, England"                            
#> [101] "Travel time, destination and origin indicators to Pharmacy by cycle and car, local authority, England"                            
#> [102] "Travel time, destination and origin indicators to Pharmacy by cycle and car, local authority, England"                            
#> [103] "Travel time, destination and origin indicators to Pharmacy by cycle and car, local authority, England"                            
#> [104] "Travel time, destination and origin indicators for Employment centres by mode of travel, Lower Super Output Area (LSOA), England "
#> [105] "Travel time, destination and origin indicators for Employment centres by mode of travel, Lower Super Output Area (LSOA), England "
#> [106] "Travel time, destination and origin indicators for Employment centres by mode of travel, Lower Super Output Area (LSOA), England "
#> [107] "Travel time, destination and origin indicators for Employment centres by mode of travel, Lower Super Output Area (LSOA), England "
#> [108] "Travel time, destination and origin indicators for Employment centres by mode of travel, Lower Super Output Area (LSOA), England "
#> [109] "Travel time, destination and origin indicators for Primary schools by mode of travel, Lower Super Output Area (LSOA), England"    
#> [110] "Travel time, destination and origin indicators for Primary schools by mode of travel, Lower Super Output Area (LSOA), England"    
#> [111] "Travel time, destination and origin indicators for Primary schools by mode of travel, Lower Super Output Area (LSOA), England"    
#> [112] "Travel time, destination and origin indicators for Primary schools by mode of travel, Lower Super Output Area (LSOA), England"    
#> [113] "Travel time, destination and origin indicators for Primary schools by mode of travel, Lower Super Output Area (LSOA), England"    
#> [114] "Travel time, destination and origin indicators for Secondary schools by mode of travel, Lower Super Output Area (LSOA), England"  
#> [115] "Travel time, destination and origin indicators for Secondary schools by mode of travel, Lower Super Output Area (LSOA), England"  
#> [116] "Travel time, destination and origin indicators for Secondary schools by mode of travel, Lower Super Output Area (LSOA), England"  
#> [117] "Travel time, destination and origin indicators for Secondary schools by mode of travel, Lower Super Output Area (LSOA), England"  
#> [118] "Travel time, destination and origin indicators for Secondary schools by mode of travel, Lower Super Output Area (LSOA), England"  
#> [119] "Travel time, destination and origin indicators for Further education by mode of travel, Lower Super Output Area (LSOA), England"  
#> [120] "Travel time, destination and origin indicators for Further education by mode of travel, Lower Super Output Area (LSOA), England"  
#> [121] "Travel time, destination and origin indicators for Further education by mode of travel, Lower Super Output Area (LSOA), England"  
#> [122] "Travel time, destination and origin indicators for Further education by mode of travel, Lower Super Output Area (LSOA), England"  
#> [123] "Travel time, destination and origin indicators for Further education by mode of travel, Lower Super Output Area (LSOA), England"  
#> [124] "Travel time, destination and origin indicators for GPs by mode of travel, Lower Super Output Area (LSOA), England"                
#> [125] "Travel time, destination and origin indicators for GPs by mode of travel, Lower Super Output Area (LSOA), England"                
#> [126] "Travel time, destination and origin indicators for GPs by mode of travel, Lower Super Output Area (LSOA), England"                
#> [127] "Travel time, destination and origin indicators for GPs by mode of travel, Lower Super Output Area (LSOA), England"                
#> [128] "Travel time, destination and origin indicators for GPs by mode of travel, Lower Super Output Area (LSOA), England"                
#> [129] "Travel time, destination and origin indicators for Hospitals by mode of travel, Lower Super Output Area (LSOA), England"          
#> [130] "Travel time, destination and origin indicators for Hospitals by mode of travel, Lower Super Output Area (LSOA), England"          
#> [131] "Travel time, destination and origin indicators for Hospitals by mode of travel, Lower Super Output Area (LSOA), England"          
#> [132] "Travel time, destination and origin indicators for Hospitals by mode of travel, Lower Super Output Area (LSOA), England"          
#> [133] "Travel time, destination and origin indicators for Hospitals by mode of travel, Lower Super Output Area (LSOA), England"          
#> [134] "Travel time, destination and origin indicators for Food stores by mode of travel, Lower Super Output Area (LSOA), England "       
#> [135] "Travel time, destination and origin indicators for Food stores by mode of travel, Lower Super Output Area (LSOA), England "       
#> [136] "Travel time, destination and origin indicators for Food stores by mode of travel, Lower Super Output Area (LSOA), England "       
#> [137] "Travel time, destination and origin indicators for Food stores by mode of travel, Lower Super Output Area (LSOA), England "       
#> [138] "Travel time, destination and origin indicators for Food stores by mode of travel, Lower Super Output Area (LSOA), England "       
#> [139] "Travel time, destination and origin indicators for town centres by mode of travel, Lower Super Output Area (LSOA), England"       
#> [140] "Travel time, destination and origin indicators for town centres by mode of travel, Lower Super Output Area (LSOA), England"       
#> [141] "Travel time, destination and origin indicators for town centres by mode of travel, Lower Super Output Area (LSOA), England"       
#> [142] "Travel time, destination and origin indicators for town centres by mode of travel, Lower Super Output Area (LSOA), England"       
#> [143] "Travel time, destination and origin indicators for town centres by mode of travel, Lower Super Output Area (LSOA), England"       
#> [144] "Travel time, destination and origin indicators to Pharmacies by cycle and car, Lower Super Output Area (LSOA), England"           
#> [145] "Travel time, destination and origin indicators to Pharmacies by cycle and car, Lower Super Output Area (LSOA), England"           
#> [146] "Travel time, destination and origin indicators to Pharmacies by cycle and car, Lower Super Output Area (LSOA), England"           
#> [147] "Travel time, destination and origin indicators to Pharmacies by cycle and car, Lower Super Output Area (LSOA), England"           
#> [148] "Travel time, destination and origin indicators to Pharmacies by cycle and car, Lower Super Output Area (LSOA), England"           
#> [149] "Average minimum journey times to nearest of selected airports by local authority: England"                                        
#> [150] "Average minimum journey times to nearest of selected airports by local authority: England"                                        
#> [151] "Average minimum journey times to nearest of selected airports by local authority: England"                                        
#> [152] "Average minimum journey times to nearest of selected airports by local authority: England"                                        
#> [153] "Average minimum journey times to nearest of selected airports by local authority: England"                                        
#> [154] "Average minimum journey times to nearest of selected airports by local authority: England"                                        
#> [155] "Catchment population measures for selected airports: England"                                                                     
#> [156] "Catchment population measures for selected airports: England"                                                                     
#> [157] "Catchment population measures for selected airports: England"                                                                     
#> [158] "Catchment population measures for selected airports: England"                                                                     
#> [159] "Catchment population measures for selected airports: England"                                                                     
#> [160] "Population within specified journey times of selected airports by local authority: England"                                       
#> [161] "Population within specified journey times of selected airports by local authority: England"                                       
#> [162] "Population within specified journey times of selected airports by local authority: England"                                       
#> [163] "Population within specified journey times of selected airports by local authority: England"                                       
#> [164] "Population within specified journey times of selected airports by local authority: England"                                       
#> [165] "Average number of selected airports within specified journey time of each area by local authority: England "                      
#> [166] "Average number of selected airports within specified journey time of each area by local authority: England "                      
#> [167] "Average number of selected airports within specified journey time of each area by local authority: England "                      
#> [168] "Average number of selected airports within specified journey time of each area by local authority: England "                      
#> [169] "Average number of selected airports within specified journey time of each area by local authority: England "                      
#> [170] "Average minimum journey times for selected airports by Lower Super Output Area (LSOA11): England"                                 
#> [171] "Average minimum journey times for selected airports by Lower Super Output Area (LSOA11): England"                                 
#> [172] "Average minimum journey times for selected airports by Lower Super Output Area (LSOA11): England"                                 
#> [173] "Average minimum journey times for selected airports by Lower Super Output Area (LSOA11): England"                                 
#> [174] "Average minimum journey times for selected airports by Lower Super Output Area (LSOA11): England"                                 
#> [175] "Average minimum journey times for selected airports by Lower Super Output Area (LSOA11): England"                                 
#> [176] "Average minimum journey times to nearest of selected rail stations by local authority: England"                                   
#> [177] "Average minimum journey times to nearest of selected rail stations by local authority: England"                                   
#> [178] "Average minimum journey times to nearest of selected rail stations by local authority: England"                                   
#> [179] "Average minimum journey times to nearest of selected rail stations by local authority: England"                                   
#> [180] "Average minimum journey times to nearest of selected rail stations by local authority: England"                                   
#> [181] "Average minimum journey times to nearest of selected rail stations by local authority: England"                                   
#> [182] "Average minimum journey times to nearest of selected rail stations by local authority: England"                                   
#> [183] "Catchment population measures for selected rail stations: England"                                                                
#> [184] "Catchment population measures for selected rail stations: England"                                                                
#> [185] "Catchment population measures for selected rail stations: England"                                                                
#> [186] "Catchment population measures for selected rail stations: England"                                                                
#> [187] "Catchment population measures for selected rail stations: England"                                                                
#> [188] "Population within specified journey times of selected rail stations by local authority: England"                                  
#> [189] "Population within specified journey times of selected rail stations by local authority: England"                                  
#> [190] "Population within specified journey times of selected rail stations by local authority: England"                                  
#> [191] "Population within specified journey times of selected rail stations by local authority: England"                                  
#> [192] "Population within specified journey times of selected rail stations by local authority: England"                                  
#> [193] "Average number of selected rail stations within specified journey time of each area by local authority: England"                  
#> [194] "Average number of selected rail stations within specified journey time of each area by local authority: England"                  
#> [195] "Average number of selected rail stations within specified journey time of each area by local authority: England"                  
#> [196] "Average number of selected rail stations within specified journey time of each area by local authority: England"                  
#> [197] "Average number of selected rail stations within specified journey time of each area by local authority: England"                  
#> [198] "Average minimum journey times by car for selected rail stations by Lower Super Output Area (LSOA11): England"                     
#> [199] "Average minimum journey times by car for selected rail stations by Lower Super Output Area (LSOA11): England"                     
#> [200] "Average minimum journey times by car for selected rail stations by Lower Super Output Area (LSOA11): England"                     
#> [201] "Average minimum journey times by car for selected rail stations by Lower Super Output Area (LSOA11): England"                     
#> [202] "Average minimum journey times by car for selected rail stations by Lower Super Output Area (LSOA11): England"                     
#> [203] "Average minimum journey times by car for selected rail stations by Lower Super Output Area (LSOA11): England"                     
#> [204] "Average minimum journey times by car for selected rail stations by Lower Super Output Area (LSOA11): England"                     
#> [205] "Average minimum journey times by car for selected rail stations by Lower Super Output Area (LSOA11): England"                     
#> [206] "Average minimum journey times by car for selected rail stations by Lower Super Output Area (LSOA11): England"                     
#> [207] "Average minimum journey times by car for selected rail stations by Lower Super Output Area (LSOA11): England"                     
#> [208] "Average minimum journey times by car for selected rail stations by Lower Super Output Area (LSOA11): England"                     
#> [209] "Average minimum journey times by car for selected rail stations by Lower Super Output Area (LSOA11): England"                     
#> [210] "Average travel time to city centres: England"                                                                                     
#> [211] "Average travel time to city centres: England"                                                                                     
#> [212] "Average travel time to city centres: England"                                                                                     
#> [213] "Average travel time to city centres: England"                                                                                     
#> [214] "Average travel time to city centres: England"                                                                                     
#> [215] "Average travel time to city centres: England"                                                                                     
#> [216] "Average travel time to city centres: England"                                                                                     
#> [217] "Average travel time to job centres in England"                                                                                    
#> [218] "Average travel time to job centres in England"                                                                                    
#> [219] "Average travel time to job centres in England"                                                                                    
#> [220] "Average travel time to job centres in England"                                                                                    
#> [221] "Average travel time to job centres in England"                                                                                    
#> [222] "Average travel time to job centres in England"                                                                                    
#> [223] "Average travel time to job centres in England"                                                                                    

get_jts_data(table = "jts0101", year = 2017, skip = 7)

#> This table's title is Average minimum travel time to reach the nearest key services by mode of travel: England

#> These data files are available for that table code: jts0101-2014.csv
#> jts0101-2015.csv
#> jts0101-2016.csv
#> jts0101-2017.csv
#> jts0101-JTS0101.csv

#> Reading in file https://github.com/cyipt/acton/releases/download/0.0.1/jts0101-2017.csv

#> Warning: Missing column names filled in: 'X15' [15], 'X16' [16]

#> Parsed with column specification:
#> cols(
#>   Year = col_character(),
#>   Mode = col_character(),
#>   `Area type 1` = col_character(),
#>   `Places with 100-499 jobs` = col_double(),
#>   `Places with 500-4999 jobs` = col_double(),
#>   `Places with 5000 or more jobs` = col_double(),
#>   `Primary school` = col_double(),
#>   `Secondary school` = col_double(),
#>   `Further Education` = col_double(),
#>   GP = col_double(),
#>   Hospital = col_double(),
#>   `Food store` = col_double(),
#>   `Town Centres` = col_double(),
#>   `Average of 8 services2` = col_character(),
#>   X15 = col_logical(),
#>   X16 = col_logical()
#> )

#> # A tibble: 47 x 14
#>    Year  Mode  `Area type 1` `Places with Jo… `Places with Jo… `Places with Jo…
#>    <chr> <chr> <chr>                    <dbl>            <dbl>            <dbl>
#>  1 2017  Publ… Urban major …              6.6              9.2             23.6
#>  2 <NA>  <NA>  Urban minor …              7.5             10.9             26.1
#>  3 <NA>  <NA>  Urban city a…              8.3             11               30.1
#>  4 <NA>  <NA>  Urban city a…              9.8              9.7             76.4
#>  5 <NA>  <NA>  Rural town a…             10.7             13.9             45.3
#>  6 <NA>  <NA>  Rural town a…             14.3             11.6             81.9
#>  7 <NA>  <NA>  Rural village             22.3             23.6             59.9
#>  8 <NA>  <NA>  Rural villag…             32.8             34.3             97.5
#>  9 <NA>  <NA>  Rural hamlet…             29.9             31               67.6
#> 10 <NA>  <NA>  Rural hamlet…             52.2             56              105. 
#> # … with 37 more rows, and 8 more variables: `Primary school` <dbl>, `Secondary
#> #   school` <dbl>, `Further Education` <dbl>, GP <dbl>, Hospital <dbl>, `Food
#> #   store` <dbl>, `Town Centres` <dbl>, `Average of 8 services2` <chr>

# uncomment on released version
# get_jts_data(table = "jts0401", year = 2017)
# get_jts_data(table = "jts0402", year = 2017)
# get_jts_data(table = "jts0402", year = 2014)