Get data from Planit API

Get data from Planit API

get_planit_data(
  bbox,
  base_url = "https://www.planit.org.uk/api/applics/json",
  limit = 6,
  end_date = "2009-02-01",
  start_date = "2000-02-01",
  silent = FALSE
)

Arguments

bbox
Bounding box in the form of xmin, ymin, xmax, ymax, e.g.: `c(-1.3, 53.7, -1.2, 53.9)`

base_url
The base URL of the service

limit
How many items to return (e.g. 6, default)

end_date
E.g. `"2009-02-01"`

start_date
E.g. `"2000-02-01"`

silent
Do you want a message? Default is `FALSE`

Value

A data frame

Examples

get_planit_data(bbox = c(-1.366023, 53.744171, -1.35515, 53.747495))

#> Getting data from https://www.planit.org.uk/api/applics/json?limit=30&bbox=-1.366023%2C53.744171%2C-1.35515%2C53.747495&end_date=2009-02-01&start_date=2000-02-01

#>     doc_type              name
#> 1 PlanApplic Leeds/08/05256/FU
#> 2 PlanApplic Leeds/08/04867/FU
#> 3 PlanApplic Leeds/08/03416/FU
#> 4 PlanApplic Leeds/07/05524/FU
#> 5 PlanApplic Leeds/07/01559/FU
#> 6 PlanApplic Leeds/06/04499/FU
#> 7 PlanApplic Leeds/06/03907/FU
#> 8 PlanApplic Leeds/06/01651/FU
#> 9 PlanApplic Leeds/06/01828/FU
#>                                                                                                                  url
#> 1 https://publicaccess.leeds.gov.uk/online-applications/applicationDetails.do?activeTab=summary&keyVal=K6Q9CIJB0FZ00
#> 2 https://publicaccess.leeds.gov.uk/online-applications/applicationDetails.do?activeTab=summary&keyVal=K5JICFJB0FZ00
#> 3 https://publicaccess.leeds.gov.uk/online-applications/applicationDetails.do?activeTab=summary&keyVal=K279JIJB0FR00
#> 4 https://publicaccess.leeds.gov.uk/online-applications/applicationDetails.do?activeTab=summary&keyVal=JNSAOWJB0FZ00
#> 5 https://publicaccess.leeds.gov.uk/online-applications/applicationDetails.do?activeTab=summary&keyVal=JEMVN0JBZF000
#> 6 https://publicaccess.leeds.gov.uk/online-applications/applicationDetails.do?activeTab=summary&keyVal=J2Y67VJB1N000
#> 7 https://publicaccess.leeds.gov.uk/online-applications/applicationDetails.do?activeTab=summary&keyVal=J1GR5AJB1N000
#> 8 https://publicaccess.leeds.gov.uk/online-applications/applicationDetails.do?activeTab=summary&keyVal=IW6172JBZF000
#> 9 https://publicaccess.leeds.gov.uk/online-applications/applicationDetails.do?activeTab=summary&keyVal=IWHID3JBZF000
#>                                                                             description
#> 1                                                          Single storey side extension
#> 2                                                          Single storey rear extension
#> 3                                   Two storey side extension and dormer window to rear
#> 4                                                                  Conservatory to rear
#> 5      Part two storey and part single storey side extension and part new roof to front
#> 6 Lower ground floor extension to side with conservatory over and terraced area to rear
#> 7                                                             Two storey side extension
#> 8                                                          Single storey rear extension
#> 9                                                          Single storey rear extension
#>                       when_updated authority_id
#> 1 2017-07-04T03:43:34.686000+01:00          292
#> 2 2017-07-05T01:51:01.397000+01:00          292
#> 3 2017-07-08T21:52:41.404000+01:00          292
#> 4 2017-09-02T13:42:29.552000+01:00          292
#> 5 2017-10-09T21:40:42.644000+01:00          292
#> 6 2019-01-02T19:55:37.177000+00:00          292
#> 7 2019-01-17T00:12:46.804000+00:00          292
#> 8 2019-03-17T09:53:18.486000+00:00          292
#> 9 2019-03-26T05:44:50.874000+00:00          292
#>                                                                        source_url
#> 1 https://publicaccess.leeds.gov.uk/online-applications/search.do?action=advanced
#> 2 https://publicaccess.leeds.gov.uk/online-applications/search.do?action=advanced
#> 3 https://publicaccess.leeds.gov.uk/online-applications/search.do?action=advanced
#> 4 https://publicaccess.leeds.gov.uk/online-applications/search.do?action=advanced
#> 5 https://publicaccess.leeds.gov.uk/online-applications/search.do?action=advanced
#> 6 https://publicaccess.leeds.gov.uk/online-applications/search.do?action=advanced
#> 7 https://publicaccess.leeds.gov.uk/online-applications/search.do?action=advanced
#> 8 https://publicaccess.leeds.gov.uk/online-applications/search.do?action=advanced
#> 9 https://publicaccess.leeds.gov.uk/online-applications/search.do?action=advanced
#>   authority_name                                                    link
#> 1          Leeds https://www.planit.org.uk/planapplic/Leeds/08/05256/FU/
#> 2          Leeds https://www.planit.org.uk/planapplic/Leeds/08/04867/FU/
#> 3          Leeds https://www.planit.org.uk/planapplic/Leeds/08/03416/FU/
#> 4          Leeds https://www.planit.org.uk/planapplic/Leeds/07/05524/FU/
#> 5          Leeds https://www.planit.org.uk/planapplic/Leeds/07/01559/FU/
#> 6          Leeds https://www.planit.org.uk/planapplic/Leeds/06/04499/FU/
#> 7          Leeds https://www.planit.org.uk/planapplic/Leeds/06/03907/FU/
#> 8          Leeds https://www.planit.org.uk/planapplic/Leeds/06/01651/FU/
#> 9          Leeds https://www.planit.org.uk/planapplic/Leeds/06/01828/FU/
#>   postcode                                                address      lat
#> 1 WF10 2AR               55 Park Avenue Allerton Bywater WF10 2AR 53.74600
#> 2 WF10 2AW  8 Barnsdale Road Allerton Bywater Castleford WF10 2AW 53.74557
#> 3 WF10 2AR    49 Park Avenue Allerton Bywater Castleford WF10 2AR 53.74600
#> 4 WF10 2AP      70 Park Lane Allerton Bywater Castleford WF10 2AP 53.74688
#> 5 WF10 2AW   1 Barnsdale Road Allerton Bywater Wakefield WF10 2AW 53.74557
#> 6 WF10 2AW 2 Junction Houses Allerton Bywater Castleford WF10 2AW 53.74557
#> 7 WF10 2AP      80 Park Lane Allerton Bywater Castleford WF10 2AP 53.74688
#> 8 WF10 2AP           78 Park Lane Allerton Bywater Leeds WF10 2AP 53.74688
#> 9 WF10 2AP           76 Park Lane Allerton Bywater Leeds WF10 2AP 53.74688
#>         lng start_date         uid
#> 1 -1.357213 2008-09-05 08/05256/FU
#> 2 -1.355825 2008-08-13 08/04867/FU
#> 3 -1.357213 2008-07-04 08/03416/FU
#> 4 -1.356836 2007-08-31 07/05524/FU
#> 5 -1.355825 2007-03-08 07/01559/FU
#> 6 -1.355825 2006-07-24 06/04499/FU
#> 7 -1.356836 2006-06-26 06/03907/FU
#> 8 -1.356836 2006-03-24 06/01651/FU
#> 9 -1.356836 2006-03-20 06/01828/FU